Handling function outputs as their own types
Hi,

With the current version of toXML, if we render an object that has a function, the values returned by the function are always treated as a string. For example, if we do:

    var person = {
        pets: function () {
            return [{ dog: "Rex" }, { cat: "Greebo" }];
        }
    }

    var xml = toXML(person);

The value of `xml` will be:

    <pets>[object Object],[object Object]</pets>

The change proposed above, the results of functions are fed again into the convert cycle, so the results of a function are treated as their own types. The output of the call to `toXML`, with the same input as above, would be:

    <pets><dog>Rex</dog><cat>Greebo</cat></pets>